(PC-17992)[API] feat: poc: new backoffice
Add some basic routes for authentication and search.
This new (POC) backoffice does work with a SPA, the whole html is
rendered by the server.

Note regarding the login manager:

Depending on context, do not load the user the same way. The default
case stays the same: load the whole User object. But for the backoffice
we also need to load the information from the backoffice_user_profile
related table. And we don not need to load all of the user's
information.